refactor: extract custom provider logic to pkg/config
- Move custom provider business logic from cmd/mpt to pkg/config/custom.go
- Create CustomProviderManager to encapsulate provider configuration
- Fix temperature=0 bug: use -1 as sentinel for unset temperature
- Fix AnyEnabled to check actual enabled status, not just configuration
- Add comprehensive test coverage for precedence merging and edge cases
- Remove todo_custom.md that shouldn't be in version control
- Update README defaults to match actual code values
